requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.23.0
scipy>=1.10.0

# Model & Training
transformers>=4.30.0
huggingface-hub>=0.16.0
datasets>=2.13.0
accelerate>=0.20.0
peft>=0.4.0

# ASR Models
openai-whisper>=20230314
# Note: Install additional models via command line as needed

# Evaluation & Metrics
jiwer>=2.6.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Demo
gradio>=3.41.0
streamlit>=1.27.0
plotly>=5.15.0

# CLI & Utilities
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
ruff>=0.0.280

# Logging & Monitoring
python-json-logger>=2.0.0

# Optional: For advanced features
pyannote.audio>=2.1.0  # Speaker diarization
moviepy>=1.0.0  # Optional for video processing
